+++ tests/Neo.L2.Abstractions.UnitTests/UT_NeoHubDeployReport.cs
@@ -199,7 +199,7 @@ public class UT_NeoHubDeployReport
             Assert.AreEqual(
                 "LocalHost.NextExpectedBlock / LastAcknowledgedBatchNumber / LastAcknowledgedBlock / "
                 + "NextBatchNumber / HasPendingSealedBatch / PendingSealedBatchNumber / "
-                + "PendingSealedBatchLastBlock / IsBatcherEnabled / PendingSealedBatch / HasOpenBatch / "
+                + "PendingSealedBatchLastBlock / IsBatcherEnabled / MaxBlocksPerBatch / MaxTransactionsPerBatch / MaxBatchAgeMillis / PendingSealedBatch / HasOpenBatch / "
                 + "InProgressTxCount / OpenBatchFirstBlock / OpenBatchLastBlock / OpenBatchBlockCount / "
                 + "OpenBatchL1MessageCount / OpenBatchL2ToL1MessageCount / ProcessCommittedBlock / "
                 + "TryRetryPendingSealedBatch / OnBatchSealed",